refactor(ytdlp): split ytDlpUtils into focused modules
Break the 2145-line utils/ytDlpUtils.ts into cohesive modules under
utils/ytdlp/ (constants, spawnEnv, versionProbe, pathResolver, install,
cookies, runtime, extractorArgs, flags, config, proxy, execute) and turn
ytDlpUtils.ts into an explicit re-export barrel.

Module-level caches now live with the concern that owns them, each
exposing a small reset* helper; install.ts and the test-reset helper
compose those resets instead of mutating shared globals. Behavior is
unchanged, including the nosemgrep annotations on every spawn call.
