+++ backend/src/utils/ytDlpUtils.ts
@@ -1,2144 +1,35 @@
-import { spawn } from "child_process";
-import fs from "fs-extra";
-import path from "path";
-import { PassThrough } from "stream";
-import { SocksProxyAgent } from "socks-proxy-agent";
-import { isAdminTrustLevelAtLeast } from "../config/adminTrust";
-import { COOKIES_FILENAME, DATA_DIR } from "../config/paths";
-import {
-  getProviderPluginPath,
-  getProviderScript,
-} from "../services/downloaders/ytdlp/ytdlpHelpers";
-import * as storageService from "../services/storageService";
-import { isBilibiliUrl, isYouTubeUrl } from "./helpers";
-import {
-  moveSafeSync,
-  pathExistsSafeSync,
-  pathExistsTrustedSync,
-  readdirSafeSync,
-  readFileSafeSync,
-  resolveSafeChildPath,
-  statSafeSync,
-  writeFileSafeSync,
-} from "./security";
-import {
-  UnsupportedCookieFormatError,
-  isValidNetscapeCookiesFile,
-  normalizeCookiesFileContent,
-} from "./cookieFileFormat";
-
-const DEFAULT_YT_DLP_PATH = "yt-dlp";
-const YT_DLP_JS_RUNTIME_ENV = "YT_DLP_JS_RUNTIME";
-const YT_DLP_HELP_PROBE_TIMEOUT_MS = 5000;
-const YT_DLP_STALE_AFTER_DAYS = 90;
-const DEFAULT_YOUTUBE_PLAYER_CLIENT_EXTRACTOR_ARG =
-  "youtube:player_client=default,mweb";
-const DEFAULT_YOUTUBE_REMOTE_COMPONENTS = "ejs:github";
-const YOUTUBE_PLAYER_CLIENT_ARG_PREFIX = "youtube:player_client=";
-const PROVIDER_SCRIPT_ARG_PREFIX = "youtubepot-bgutilscript:script_path=";
-
-// Cached promise so we only check/install once per process
-let ytDlpAvailablePromise: Promise<void> | null = null;
-let denoAvailablePromise: Promise<boolean> | null = null;
-let providerPluginPathCache: string | null | undefined = undefined;
-let resolvedYtDlpPathCache: string | null = null;
-type YouTubeJsRuntimeFlag = "--js-runtime" | "--js-runtimes";
-type CookiesFileSignature = {
-  mtimeMs: number;
-  size: number;
-};
-type CookiesFileCache = CookiesFileSignature & {
-  path: string | null;
-};
-
-let jsRuntimeFlagPromise: Promise<YouTubeJsRuntimeFlag | null> | null = null;
-let remoteComponentsSupportPromise: Promise<boolean> | null = null;
-let impersonateSupportPromise: Promise<boolean> | null = null;
-const runtimeWarningCache = new Set<string>();
-let cookiesFileCache: CookiesFileCache | null = null;
-
-function appendUniquePathEntry(entries: string[], candidate: string): void {
-  if (!candidate || entries.includes(candidate)) {
-    return;
-  }
-
-  entries.push(candidate);
-}
-
-function prependUniquePathEntry(entries: string[], candidate: string): void {
-  if (!candidate) {
-    return;
-  }
-
-  const existingIndex = entries.indexOf(candidate);
-  if (existingIndex === 0) {
-    return;
-  }
-
-  if (existingIndex > 0) {
-    entries.splice(existingIndex, 1);
-  }
-
-  entries.unshift(candidate);
-}
-
-function appendSafeChildDir(
-  entries: string[],
-  allowedDir: string,
-  childDir: string,
-): void {
-  try {
-    appendUniquePathEntry(entries, resolveSafeChildPath(allowedDir, childDir));
-  } catch {
-    // Ignore invalid child paths discovered while scanning local install roots.
-  }
-}
-
-function getYtDlpExecutableNames(): string[] {
-  return process.platform === "win32"
-    ? ["yt-dlp.exe", "yt-dlp.cmd", "yt-dlp.bat", "yt-dlp"]
-    : ["yt-dlp"];
-}
-
-function appendWindowsPythonScriptsDirs(
-  candidateDirs: string[],
-  rootDir: string,
-): void {
-  try {
-    for (const entry of readdirSafeSync(rootDir, rootDir)) {
-      if (!/^python\d+(?:-\d+)?$/i.test(entry)) {
-        continue;
-      }
-      const pythonDir = resolveSafeChildPath(rootDir, entry);
-      appendSafeChildDir(candidateDirs, pythonDir, "Scripts");
-    }
-  } catch {
-    // Ignore missing or unreadable Python installation roots.
-  }
-}
-
-function listLikelyUserBinDirs(): string[] {
-  const candidateDirs: string[] = [];
-
-  if (process.platform === "win32") {
-    const userProfile = process.env.USERPROFILE?.trim();
-    const appData =
-      process.env.APPDATA?.trim() ||
-      (userProfile ? path.join(userProfile, "AppData", "Roaming") : "");
-    const localAppData =
-      process.env.LOCALAPPDATA?.trim() ||
-      (userProfile ? path.join(userProfile, "AppData", "Local") : "");
-
-    if (appData) {
-      appendUniquePathEntry(candidateDirs, path.join(appData, "Python", "Scripts"));
-      appendWindowsPythonScriptsDirs(candidateDirs, path.join(appData, "Python"));
-    }
-
-    if (localAppData) {
-      appendWindowsPythonScriptsDirs(
-        candidateDirs,
-        path.join(localAppData, "Programs", "Python")
-      );
-    }
-
-    return candidateDirs;
-  }
-
-  const homeDir = process.env.HOME?.trim();
-  if (!homeDir) {
-    return candidateDirs;
-  }
-
-  appendUniquePathEntry(candidateDirs, path.join(homeDir, ".local", "bin"));
-
-  if (process.platform === "darwin") {
-    const macUserPythonRoot = path.join(homeDir, "Library", "Python");
-    try {
-      for (const entry of readdirSafeSync(macUserPythonRoot, macUserPythonRoot)) {
-        if (!/^\d+\.\d+$/.test(entry)) {
-          continue;
-        }
-        const pythonDir = resolveSafeChildPath(macUserPythonRoot, entry);
-        appendSafeChildDir(candidateDirs, pythonDir, "bin");
-      }
-    } catch {
-      // Ignore missing or unreadable user Python directories.
-    }
-  }
-
-  return candidateDirs;
-}
-
-function resolveCookiesPath(): string {
-  return resolveSafeChildPath(DATA_DIR, COOKIES_FILENAME);
-}
-
-function getCachedProviderPluginPath(): string {
-  if (providerPluginPathCache === undefined) {
-    providerPluginPathCache = getProviderPluginPath() || null;
-  }
-
-  return providerPluginPathCache || "";
-}
-
-function getExplicitConfiguredYtDlpPath(): string | null {
-  const configuredPath = process.env.YT_DLP_PATH?.trim();
-  return configuredPath ? configuredPath : null;
-}
-
-function getConfiguredYtDlpPath(): string {
-  return (
-    getExplicitConfiguredYtDlpPath() ||
-    resolvedYtDlpPathCache ||
-    DEFAULT_YT_DLP_PATH
-  );
-}
-
-function getYtDlpSpawnEnv(): NodeJS.ProcessEnv {
-  const providerPluginPath = getCachedProviderPluginPath();
-  if (!providerPluginPath) {
-    return process.env;
-  }
-
-  const pythonPathEntries = (process.env.PYTHONPATH ?? "")
-    .split(path.delimiter)
-    .filter(Boolean);
-
-  if (!pythonPathEntries.includes(providerPluginPath)) {
-    pythonPathEntries.unshift(providerPluginPath);
-  }
-
-  return {
-    ...process.env,
-    PYTHONPATH: pythonPathEntries.join(path.delimiter),
-  };
-}
-
-function hasCustomConfiguredYtDlpPath(): boolean {
-  return Boolean(
-    getExplicitConfiguredYtDlpPath() &&
-      getExplicitConfiguredYtDlpPath() !== DEFAULT_YT_DLP_PATH
-  );
-}
-
-function listYtDlpPathCandidates(): string[] {
-  const pathEntries = (process.env.PATH ?? "")
-    .split(path.delimiter)
-    .filter(Boolean);
-  for (const candidateDir of listLikelyUserBinDirs()) {
-    appendUniquePathEntry(pathEntries, candidateDir);
-  }
-  const executableNames = getYtDlpExecutableNames();
-
-  const seen = new Set<string>();
-  const candidates: string[] = [];
-
-  for (const entry of pathEntries) {
-    for (const executableName of executableNames) {
-      let candidatePath: string;
-      try {
-        candidatePath = resolveSafeChildPath(entry, executableName);
-      } catch {
-        continue;
-      }
-      let candidateExists = false;
-      try {
-        candidateExists = pathExistsSafeSync(candidatePath, entry);
-      } catch {
-        continue;
-      }
-      if (!candidateExists || seen.has(candidatePath)) {
-        continue;
-      }
-      seen.add(candidatePath);
-      candidates.push(candidatePath);
-    }
-  }
-
-  return candidates;
-}
-
-function updatePathAfterAutoInstall(): void {
-  const pathEntries = (process.env.PATH ?? "")
-    .split(path.delimiter)
-    .filter(Boolean);
-
-  for (const candidateDir of listLikelyUserBinDirs()) {
-    const hasExecutable = getYtDlpExecutableNames().some((executableName) => {
-      try {
-        const candidatePath = `${candidateDir}${path.sep}${executableName}`;
-        return pathExistsTrustedSync(candidatePath);
-      } catch {
-        return false;
-      }
-    });
-
-    if (hasExecutable) {
-      prependUniquePathEntry(pathEntries, candidateDir);
-    }
-  }
-
-  process.env.PATH = pathEntries.join(path.delimiter);
-}
-
-type YtDlpCandidateProbe = {
-  canRun: boolean;
-  supportsModernJsRuntimes: boolean;
-  version: string | null;
-  releaseTimestamp: number | null;
-  isStale: boolean;
-};
-
-function extractYtDlpReleaseTimestamp(versionText: string | null): number | null {
-  if (!versionText) {
-    return null;
-  }
-
-  const match = versionText.match(/(\d{4})\.(\d{2})\.(\d{2})/);
-  if (!match) {
-    return null;
-  }
-
-  const year = Number(match[1]);
-  const monthIndex = Number(match[2]) - 1;
-  const day = Number(match[3]);
-  const timestamp = Date.UTC(year, monthIndex, day);
-  return Number.isFinite(timestamp) ? timestamp : null;
-}
-
-function isYtDlpReleaseStale(releaseTimestamp: number | null): boolean {
-  if (releaseTimestamp === null) {
-    return false;
-  }
-
-  const staleAfterMs = YT_DLP_STALE_AFTER_DAYS * 24 * 60 * 60 * 1000;
-  return Date.now() - releaseTimestamp > staleAfterMs;
-}
-
-async function getYtDlpVersionInfo(ytDlpPath: string): Promise<{
-  canRun: boolean;
-  version: string | null;
-  releaseTimestamp: number | null;
-  isStale: boolean;
-  errorKind?: "close" | "spawn";
-  errorCode?: string;
-  errorMessage?: string;
-}> {
-  let versionText = "";
-
-  return await new Promise((resolve) => {
-    let settled = false;
-    const timeout = setTimeout(() => {
-      if (settled) return;
-      proc.kill("SIGTERM");
-      settled = true;
-      resolve({
-        canRun: false,
-        version: null,
-        releaseTimestamp: null,
-        isStale: false,
-        errorKind: "spawn",
-        errorCode: "ETIMEDOUT",
-        errorMessage: "yt-dlp version probe timed out",
-      });
-    }, YT_DLP_HELP_PROBE_TIMEOUT_MS);
-    timeout.unref?.();
-    const settle = (result: {
-      canRun: boolean;
-      version: string | null;
-      releaseTimestamp: number | null;
-      isStale: boolean;
-      errorKind?: "close" | "spawn";
-      errorCode?: string;
-      errorMessage?: string;
-    }) => {
-      if (settled) return;
-      settled = true;
-      clearTimeout(timeout);
-      resolve(result);
-    };
-
-    // ytDlpPath is either explicitly configured by the operator or selected
-    // from enumerated PATH entries and fixed executable names above.
-    // nosemgrep: javascript.lang.security.detect-child-process.detect-child-process
-    const proc = spawn(ytDlpPath, ["--version"], {
-      env: getYtDlpSpawnEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-
-    proc.stdout?.on("data", (data: Buffer) => {
-      versionText += data.toString();
-    });
-
-    proc.stderr?.on("data", (data: Buffer) => {
-      versionText += data.toString();
-    });
-
-    proc.on("close", (code) => {
-      if (code !== 0) {
-        settle({
-          canRun: false,
-          version: null,
-          releaseTimestamp: null,
-          isStale: false,
-          errorKind: "close",
-        });
-        return;
-      }
-
-      const normalizedVersion = versionText.trim() || null;
-      const releaseTimestamp = extractYtDlpReleaseTimestamp(normalizedVersion);
-      settle({
-        canRun: true,
-        version: normalizedVersion,
-        releaseTimestamp,
-        isStale: isYtDlpReleaseStale(releaseTimestamp),
-      });
-    });
-
-    proc.on("error", (error: NodeJS.ErrnoException) => {
-      settle({
-        canRun: false,
-        version: null,
-        releaseTimestamp: null,
-        isStale: false,
-        errorKind: "spawn",
-        errorCode: error?.code,
-        errorMessage: error?.message,
-      });
-    });
-  });
-}
-
-async function probeYtDlpCandidate(
-  ytDlpPath: string
-): Promise<YtDlpCandidateProbe> {
-  const versionInfo = await getYtDlpVersionInfo(ytDlpPath);
-  if (!versionInfo.canRun) {
-    return {
-      canRun: false,
-      supportsModernJsRuntimes: false,
-      version: null,
-      releaseTimestamp: null,
-      isStale: false,
-    };
-  }
-
-  let helpText = "";
-
-  return await new Promise<YtDlpCandidateProbe>((resolve) => {
-    let settled = false;
-    const failProbe = (): YtDlpCandidateProbe => ({
-      canRun: false,
-      supportsModernJsRuntimes: false,
-      version: versionInfo.version,
-      releaseTimestamp: versionInfo.releaseTimestamp,
-      isStale: versionInfo.isStale,
-    });
-    const timeout = setTimeout(() => {
-      if (settled) return;
-      proc.kill("SIGTERM");
-      settled = true;
-      resolve(failProbe());
-    }, YT_DLP_HELP_PROBE_TIMEOUT_MS);
-    timeout.unref?.();
-    const settle = (probe: YtDlpCandidateProbe) => {
-      if (settled) return;
-      settled = true;
-      clearTimeout(timeout);
-      resolve(probe);
-    };
-
-    // ytDlpPath is built from PATH entries and a fixed executable name above;
-    // shell execution is disabled and arguments are passed separately.
-    // nosemgrep: javascript.lang.security.detect-child-process.detect-child-process
-    const proc = spawn(ytDlpPath, ["--help"], {
-      env: getYtDlpSpawnEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-
-    proc.stdout?.on("data", (data: Buffer) => {
-      helpText += data.toString();
-    });
-
-    proc.stderr?.on("data", (data: Buffer) => {
-      helpText += data.toString();
-    });
-
-    proc.on("close", (code) => {
-      if (code !== 0) {
-        settle(failProbe());
-        return;
-      }
-      settle({
-        canRun: true,
-        supportsModernJsRuntimes:
-          helpText.includes("--js-runtimes") ||
-          helpText.includes("--js-runtime"),
-        version: versionInfo.version,
-        releaseTimestamp: versionInfo.releaseTimestamp,
-        isStale: versionInfo.isStale,
-      });
-    });
-
-    proc.on("error", () => {
-      settle(failProbe());
-    });
-  });
-}
-
-function isBetterYtDlpCandidate(
-  candidate: YtDlpCandidateProbe,
-  currentBest: YtDlpCandidateProbe | null,
-): boolean {
-  if (!currentBest) {
-    return true;
-  }
-
-  if (candidate.isStale !== currentBest.isStale) {
-    return currentBest.isStale;
-  }
-
-  if (candidate.supportsModernJsRuntimes !== currentBest.supportsModernJsRuntimes) {
-    return candidate.supportsModernJsRuntimes;
-  }
-
-  const candidateTimestamp = candidate.releaseTimestamp ?? 0;
-  const bestTimestamp = currentBest.releaseTimestamp ?? 0;
-  if (candidateTimestamp !== bestTimestamp) {
-    return candidateTimestamp > bestTimestamp;
-  }
-
-  return false;
-}
-
-async function resolveYtDlpPath(): Promise<string> {
-  const explicitPath = getExplicitConfiguredYtDlpPath();
-  if (explicitPath) {
-    return explicitPath;
-  }
-
-  if (resolvedYtDlpPathCache) {
-    return resolvedYtDlpPathCache;
-  }
-
-  const candidates = listYtDlpPathCandidates();
-  if (candidates.length === 0) {
-    resolvedYtDlpPathCache = DEFAULT_YT_DLP_PATH;
-    return resolvedYtDlpPathCache;
-  }
-
-  let bestCandidate: string | null = null;
-  let bestProbe: YtDlpCandidateProbe | null = null;
-  for (const candidate of candidates) {
-    const probe = await probeYtDlpCandidate(candidate);
-    if (!probe.canRun) {
-      continue;
-    }
-
-    if (isBetterYtDlpCandidate(probe, bestProbe)) {
-      bestCandidate = candidate;
-      bestProbe = probe;
-    }
-  }
-
-  resolvedYtDlpPathCache = bestCandidate || DEFAULT_YT_DLP_PATH;
-  return resolvedYtDlpPathCache;
-}
-
-function isMissingFileError(error: unknown): boolean {
-  return (
-    typeof error === "object" &&
-    error !== null &&
-    "code" in error &&
-    (error as NodeJS.ErrnoException).code === "ENOENT"
-  );
-}
+// Barrel for the yt-dlp utilities, split into focused modules under ./ytdlp.
+// Re-exports the public surface so existing `../utils/ytDlpUtils` imports keep working.
+
+import { resetYtDlpAvailablePromise } from "./ytdlp/install";
+import { resetProviderPluginCache } from "./ytdlp/spawnEnv";
+import { resetResolvedYtDlpPath } from "./ytdlp/pathResolver";
+import { resetRuntimeCaches } from "./ytdlp/runtime";
+import { resetCookiesFileCache } from "./ytdlp/cookies";
+
+export { ensureYtDlpAvailable } from "./ytdlp/install";
+export { isYtDlpImpersonateAvailable } from "./ytdlp/runtime";
+export { convertFlagToArg, flagsToArgs } from "./ytdlp/flags";
+export {
+  executeYtDlpJson,
+  getChannelUrlFromVideo,
+  downloadChannelAvatar,
+  executeYtDlpSpawn,
+} from "./ytdlp/execute";
+export {
+  parseYtDlpConfig,
+  getUserYtDlpConfig,
+  getNetworkConfigFromUserConfig,
+} from "./ytdlp/config";
+export { InvalidProxyError, getAxiosProxyConfig } from "./ytdlp/proxy";
 
 /**
  * @internal Test helper to reset internal caches between test cases.
  */
 export function resetYtDlpAvailabilityCacheForTests(): void {
-  ytDlpAvailablePromise = null;
-  denoAvailablePromise = null;
-  providerPluginPathCache = undefined;
-  resolvedYtDlpPathCache = null;
-  jsRuntimeFlagPromise = null;
-  remoteComponentsSupportPromise = null;
-  impersonateSupportPromise = null;
-  runtimeWarningCache.clear();
-  cookiesFileCache = null;
-}
-
-/**
- * Try to install yt-dlp via pip, trying multiple pip variants.
- */
-async function installYtDlp(options: { upgrade?: boolean } = {}): Promise<void> {
-  const { upgrade = false } = options;
-  // curl-cffi powers yt-dlp's browser impersonation (--impersonate), which the
-  // MissAV downloader needs to get past Cloudflare on the m3u8 CDN. The Docker
-  // image installs it explicitly; install it here too so non-Docker setups
-  // gain the capability instead of hard-failing on --impersonate.
-  const packages = ["yt-dlp", "bgutil-ytdlp-pot-provider", "curl-cffi"];
-  const installArgs = ["install"];
-  if (upgrade) {
-    installArgs.push("-U");
-  }
-
-  // pip candidates to try in order
-  const candidates =
-    process.platform === "win32"
-      ? [
-          ["py", "-m", "pip", ...installArgs, ...packages],
-          ["python", "-m", "pip", ...installArgs, ...packages],
-          ["python3", "-m", "pip", ...installArgs, ...packages],
-          ["pip", ...installArgs, ...packages],
-          ["pip3", ...installArgs, ...packages],
-        ]
-      : [
-          ["pip3", ...installArgs, "--user", ...packages],
-          ["pip3", ...installArgs, ...packages],
-          ["pip3", ...installArgs, "--break-system-packages", ...packages],
-          ["pip", ...installArgs, "--user", ...packages],
-          ["pip", ...installArgs, ...packages],
-          ["pip", ...installArgs, "--break-system-packages", ...packages],
-          ["python3", "-m", "pip", ...installArgs, "--user", ...packages],
-          ["python3", "-m", "pip", ...installArgs, ...packages],
-        ];
-
-  for (const [cmd, ...args] of candidates) {
-    try {
-      console.log(`[yt-dlp] Attempting: ${cmd} ${args.join(" ")}`);
-      await new Promise<void>((resolve, reject) => {
-        let stderr = "";
-        const proc = spawn(cmd, args, {
-          stdio: ["ignore", "ignore", "pipe"],
-        });
-        proc.stderr?.on("data", (data: Buffer) => {
-          stderr += data.toString();
-        });
-        proc.on("close", (code) =>
-          code === 0
-            ? resolve()
-            : reject(
-                Object.assign(new Error(`${cmd} exited with code ${code}`), {
-                  code,
-                  stderr,
-                })
-              )
-        );
-        proc.on("error", reject);
-      });
-      console.log(
-        upgrade
-          ? "[yt-dlp] Successfully updated yt-dlp and bgutil provider."
-          : "[yt-dlp] Successfully installed yt-dlp and bgutil provider."
-      );
-      updatePathAfterAutoInstall();
-      return;
-    } catch (error: any) {
-      const stderr = String(error?.stderr || "").trim();
-      if (stderr) {
-        console.warn(`[yt-dlp] ${cmd} failed: ${stderr.split("\n").pop()}`);
-      }
-      // Try next candidate
-    }
-  }
-
-  throw new Error(
-    `yt-dlp could not be automatically ${upgrade ? "updated" : "installed"}. ` +
-      "Please install it manually: https://github.com/yt-dlp/yt-dlp#installation"
-  );
-}
-
-/**
- * Ensure yt-dlp is available, auto-installing via pip if not found.
- * Result is cached so the check only runs once per process.
- */
-export async function ensureYtDlpAvailable(): Promise<void> {
-  if (ytDlpAvailablePromise) return ytDlpAvailablePromise;
-
-  ytDlpAvailablePromise = (async () => {
-    let attemptedAutoUpgrade = false;
-    let attemptedAutoInstall = false;
-
-    while (true) {
-      const ytDlpPath = await resolveYtDlpPath();
-      try {
-        const versionInfo = await getYtDlpVersionInfo(ytDlpPath);
-        if (!versionInfo.canRun) {
-          throw Object.assign(
-            new Error(versionInfo.errorMessage || "yt-dlp failed version probe"),
-            {
-              kind: versionInfo.errorKind || "close",
-              code: versionInfo.errorCode,
-            }
-          );
-        }
-
-        if (
-          !hasCustomConfiguredYtDlpPath() &&
-          !attemptedAutoUpgrade &&
-          versionInfo.isStale
-        ) {
-          attemptedAutoUpgrade = true;
-          console.warn(
-            `[yt-dlp] ${versionInfo.version || ytDlpPath} is older than ${YT_DLP_STALE_AFTER_DAYS} days. Updating yt-dlp and the bgutil provider to avoid known YouTube 360p regressions.`
-          );
-          try {
-            await installYtDlp({ upgrade: true });
-            resolvedYtDlpPathCache = null;
-            jsRuntimeFlagPromise = null;
-            remoteComponentsSupportPromise = null;
-            continue;
-          } catch (upgradeError: any) {
-            console.warn(
-              `[yt-dlp] Automatic update failed (${upgradeError?.message || "unknown error"}). Continuing with the existing yt-dlp binary.`
-            );
-            return;
-          }
-        }
-
-        return;
-      } catch (err: any) {
-        // Non-zero exit from --version means binary executed; continue.
-        if (err.kind === "close") {
-          return;
-        }
-
-        if (err.code === "EACCES" || err.code === "EPERM") {
-          throw new Error(
-            `yt-dlp exists but is not executable at: ${ytDlpPath}. ` +
-              "Please fix file permissions or install yt-dlp manually."
-          );
-        }
-
-        if (err.code === "ENOENT") {
-          // Only auto-install when using the default path (not a user-configured path).
-          if (hasCustomConfiguredYtDlpPath()) {
-            throw new Error(
-              `yt-dlp not found at configured path: ${ytDlpPath}. ` +
-                "Please check your YT_DLP_PATH environment variable."
-            );
-          }
-
-          if (attemptedAutoInstall) {
-            throw new Error(
-              "yt-dlp was installed automatically but is still not available on PATH. " +
-                "Please add it to PATH or set YT_DLP_PATH to the installed binary."
-            );
-          }
-
-          console.warn(
-            "[yt-dlp] yt-dlp not found in PATH. Attempting automatic installation..."
-          );
-          attemptedAutoInstall = true;
-          await installYtDlp();
-          resolvedYtDlpPathCache = null;
-          jsRuntimeFlagPromise = null;
-          remoteComponentsSupportPromise = null;
-          continue;
-        }
-
-        if (hasCustomConfiguredYtDlpPath()) {
-          throw new Error(
-            `Failed to execute configured yt-dlp at ${ytDlpPath} ` +
-              `(${err.code || "unknown"}): ${err.message}`
-          );
-        }
-        throw new Error(
-          `Failed to execute yt-dlp (${err.code || "unknown"}): ${err.message}`
-        );
-      }
-    }
-  })().catch((err) => {
-    // Reset cache so the next call retries instead of getting the same error.
-    ytDlpAvailablePromise = null;
-    throw err;
-  });
-
-  return ytDlpAvailablePromise;
-}
-
-/**
- * Preprocess URL to handle specific domain replacements
- * e.g. xvideos.red -> xvideos.com to support yt-dlp extraction
- */
-function preprocessUrl(url: string): string {
-  if (!url) return url;
-  
-  // Handle XVideos mirrors
-  if (url.includes("xvideos.red")) {
-    console.log(`Preprocessing URL: replacing xvideos.red with xvideos.com`);
-    return url.replace("xvideos.red", "xvideos.com");
-  }
-  
-  return url;
-}
-
-function updateCookiesFileCacheFromDisk(
-  cookiesPath: string,
-  pathValue: string | null
-): void {
-  try {
-    const stats = statSafeSync(cookiesPath, DATA_DIR);
-    cookiesFileCache = {
-      mtimeMs: stats.mtimeMs,
-      size: stats.size,
-      path: pathValue,
-    };
-  } catch {
-    cookiesFileCache = null;
-  }
-}
-
-/**
- * Return the cookies file path after ensuring the file is usable by yt-dlp.
- * Existing Cookie header files are converted in place for backward compatibility.
- */
-function ensureCookiesFileIsNormalized(): string | null {
-  const cookiesPath = resolveCookiesPath();
-  let signature: CookiesFileSignature | null = null;
-  try {
-    const stats = statSafeSync(cookiesPath, DATA_DIR);
-    signature = { mtimeMs: stats.mtimeMs, size: stats.size };
-    if (
-      cookiesFileCache &&
-      cookiesFileCache.mtimeMs === signature.mtimeMs &&
-      cookiesFileCache.size === signature.size
-    ) {
-      return cookiesFileCache.path;
-    }
-
-    const content = readFileSafeSync(cookiesPath, DATA_DIR, "utf8");
-    if (isValidNetscapeCookiesFile(content)) {
-      cookiesFileCache = { ...signature, path: cookiesPath };
-      return cookiesPath;
-    }
-
-    const normalizedContent = normalizeCookiesFileContent(content);
-    writeFileSafeSync(cookiesPath, DATA_DIR, normalizedContent, "utf8");
-    updateCookiesFileCacheFromDisk(cookiesPath, cookiesPath);
-    console.warn(
-      "[yt-dlp] Converted cookies.txt from Cookie header format to Netscape format."
-    );
-    return cookiesPath;
-  } catch (error) {
-    if (error instanceof UnsupportedCookieFormatError) {
-      if (signature) {
-        // path=null means this exact file signature is known to be unparseable.
-        cookiesFileCache = { ...signature, path: null };
-      }
-      console.warn(
-        `[yt-dlp] Ignoring invalid cookies.txt: ${error.message}`
-      );
-      return null;
-    }
-
-    if (isMissingFileError(error)) {
-      cookiesFileCache = null;
-      return null;
-    }
-
-    cookiesFileCache = null;
-    console.warn(
-      "[yt-dlp] Unable to read cookies.txt; continuing without cookies.",
-      error
-    );
-    return null;
-  }
-}
-
-async function isDenoAvailable(): Promise<boolean> {
-  if (denoAvailablePromise) {
-    return denoAvailablePromise;
-  }
-
-  denoAvailablePromise = new Promise<boolean>((resolve) => {
-    const proc = spawn("deno", ["--version"], {
-      stdio: ["ignore", "ignore", "ignore"],
-    });
-
-    proc.on("close", (code) => {
-      resolve(code === 0);
-    });
-
-    proc.on("error", () => {
-      resolve(false);
-    });
-  });
-
-  return denoAvailablePromise;
-}
-
-/**
- * Whether the resolved yt-dlp binary can impersonate a Chrome browser via
- * curl_cffi. Probes `--list-impersonate-targets`: when curl_cffi is missing,
- * yt-dlp still lists targets but tags every one "(unavailable)", and passing
- * `--impersonate` would then hard-fail the download. Callers use this to gate
- * the flag and degrade gracefully instead of erroring. Cached per process.
- */
-export async function isYtDlpImpersonateAvailable(): Promise<boolean> {
-  if (impersonateSupportPromise) {
-    return impersonateSupportPromise;
-  }
-
-  impersonateSupportPromise = (async () => {
-    try {
-      const ytDlpPath = await resolveYtDlpPath();
-      return await new Promise<boolean>((resolve) => {
-        let output = "";
-        let settled = false;
-        const finish = (value: boolean) => {
-          if (settled) return;
-          settled = true;
-          clearTimeout(timeout);
-          // A usable target row names a client and ends with the curl_cffi
-          // source without the "(unavailable)" marker.
-          const available = output
-            .split("\n")
-            .some(
-              (line) =>
-                /chrome/i.test(line) &&
-                /curl_cffi/.test(line) &&
-                !/unavailable/i.test(line),
-            );
-          if (!available) {
-            impersonateSupportPromise = null;
-          }
-          resolve(available && value);
-        };
-        const timeout = setTimeout(() => {
-          proc.kill("SIGTERM");
-          finish(true);
-        }, YT_DLP_HELP_PROBE_TIMEOUT_MS);
-        timeout.unref?.();
-
-        // ytDlpPath is either explicitly configured by the operator or selected
-        // from enumerated PATH entries and fixed executable names above.
-        // nosemgrep: javascript.lang.security.detect-child-process.detect-child-process
-        const proc = spawn(ytDlpPath, ["--list-impersonate-targets"], {
-          env: getYtDlpSpawnEnv(),
-          stdio: ["ignore", "pipe", "pipe"],
-        });
-        proc.stdout?.on("data", (data: Buffer) => {
-          output += data.toString();
-        });
-        proc.on("close", () => finish(true));
-        proc.on("error", () => {
-          impersonateSupportPromise = null;
-          finish(false);
-        });
-      });
-    } catch {
-      impersonateSupportPromise = null;
-      return false;
-    }
-  })();
-
-  return impersonateSupportPromise;
-}
-
-function warnRuntimeOnce(key: string, message: string): void {
-  if (runtimeWarningCache.has(key)) {
-    return;
-  }
-  runtimeWarningCache.add(key);
-  console.warn(message);
-}
-
-async function getYouTubeJsRuntimeFlag(): Promise<YouTubeJsRuntimeFlag | null> {
-  if (jsRuntimeFlagPromise) {
-    return jsRuntimeFlagPromise;
-  }
-
-  jsRuntimeFlagPromise = (async () => {
-    let helpText = "";
-    const resolveFromHelp = (): YouTubeJsRuntimeFlag | null => {
-      if (helpText.includes("--js-runtimes")) {
-        return "--js-runtimes";
-      }
-
-      if (helpText.includes("--js-runtime")) {
-        return "--js-runtime";
-      }
-
-      warnRuntimeOnce(
-        "js-runtime-flag-unsupported",
-        "[yt-dlp] Current yt-dlp binary does not support --js-runtimes. Continuing without it. Upgrade yt-dlp or set YT_DLP_PATH to a newer binary if YouTube extraction becomes unreliable."
-      );
-      return null;
-    };
-
-    try {
-      const ytDlpPath = await resolveYtDlpPath();
-      const resolvedFlag = await new Promise<YouTubeJsRuntimeFlag | null>((resolve) => {
-        let settled = false;
-        const timeout = setTimeout(() => {
-          if (settled) return;
-          proc.kill("SIGTERM");
-          settled = true;
-          resolve(resolveFromHelp());
-        }, YT_DLP_HELP_PROBE_TIMEOUT_MS);
-        timeout.unref?.();
-        const settle = () => {
-          if (settled) return;
-          settled = true;
-          clearTimeout(timeout);
-          resolve(resolveFromHelp());
-        };
-
-        // ytDlpPath is either explicitly configured by the operator or selected
-        // from enumerated PATH entries and fixed executable names above.
-        // nosemgrep: javascript.lang.security.detect-child-process.detect-child-process
-        const proc = spawn(ytDlpPath, ["--help"], {
-          env: getYtDlpSpawnEnv(),
-          stdio: ["ignore", "pipe", "pipe"],
-        });
-
-        proc.stdout?.on("data", (data: Buffer) => {
-          helpText += data.toString();
-        });
-
-        proc.stderr?.on("data", (data: Buffer) => {
-          helpText += data.toString();
-        });
-
-        proc.on("close", () => {
-          settle();
-        });
-
-        proc.on("error", () => {
-          settle();
-        });
-      });
-      if (!resolvedFlag) {
-        jsRuntimeFlagPromise = null;
-      }
-      return resolvedFlag;
-    } catch {
-      jsRuntimeFlagPromise = null;
-      warnRuntimeOnce(
-        "js-runtime-flag-unsupported",
-        "[yt-dlp] Current yt-dlp binary does not support --js-runtimes. Continuing without it. Upgrade yt-dlp or set YT_DLP_PATH to a newer binary if YouTube extraction becomes unreliable."
-      );
-      return null;
-    }
-  })();
-
-  return jsRuntimeFlagPromise;
-}
-
-async function ytDlpSupportsRemoteComponents(): Promise<boolean> {
-  if (remoteComponentsSupportPromise) {
-    return remoteComponentsSupportPromise;
-  }
-
-  remoteComponentsSupportPromise = (async () => {
-    let helpText = "";
-    const resolveFromHelp = (): boolean => {
-      if (helpText.includes("--remote-components")) {
-        return true;
-      }
-
-      warnRuntimeOnce(
-        "remote-components-unsupported",
-        "[yt-dlp] Current yt-dlp binary does not support --remote-components. Continuing without it. Upgrade yt-dlp or set YT_DLP_PATH to a newer binary if YouTube extraction becomes unreliable."
-      );
-      return false;
-    };
-
-    try {
-      const ytDlpPath = await resolveYtDlpPath();
-      const supported = await new Promise<boolean>((resolve) => {
-        let settled = false;
-        const timeout = setTimeout(() => {
-          if (settled) return;
-          proc.kill("SIGTERM");
-          settled = true;
-          resolve(resolveFromHelp());
-        }, YT_DLP_HELP_PROBE_TIMEOUT_MS);
-        timeout.unref?.();
-        const settle = () => {
-          if (settled) return;
-          settled = true;
-          clearTimeout(timeout);
-          resolve(resolveFromHelp());
-        };
-
-        // ytDlpPath is either explicitly configured by the operator or selected
-        // from enumerated PATH entries and fixed executable names above.
-        // nosemgrep: javascript.lang.security.detect-child-process.detect-child-process
-        const proc = spawn(ytDlpPath, ["--help"], {
-          env: getYtDlpSpawnEnv(),
-          stdio: ["ignore", "pipe", "pipe"],
-        });
-
-        proc.stdout?.on("data", (data: Buffer) => {
-          helpText += data.toString();
-        });
-
-        proc.stderr?.on("data", (data: Buffer) => {
-          helpText += data.toString();
-        });
-
-        proc.on("close", () => {
-          settle();
-        });
-
-        proc.on("error", () => {
-          settle();
-        });
-      });
-      if (!supported) {
-        remoteComponentsSupportPromise = null;
-      }
-      return supported;
-    } catch {
-      remoteComponentsSupportPromise = null;
-      warnRuntimeOnce(
-        "remote-components-unsupported",
-        "[yt-dlp] Current yt-dlp binary does not support --remote-components. Continuing without it. Upgrade yt-dlp or set YT_DLP_PATH to a newer binary if YouTube extraction becomes unreliable."
-      );
-      return false;
-    }
-  })();
-
-  return remoteComponentsSupportPromise;
-}
-
-async function getYouTubeJsRuntime(): Promise<"node" | "deno"> {
-  const rawRuntime = process.env[YT_DLP_JS_RUNTIME_ENV]?.trim();
-  const runtime = rawRuntime?.toLowerCase();
-  const hasRuntimeOverride = Boolean(rawRuntime);
-
-  if (runtime === "node") {
-    return "node";
-  }
-
-  const runtimeIsInvalid = hasRuntimeOverride && runtime !== "deno";
-
-  // Default to Deno because yt-dlp recommends it for JS challenge solving.
-  // If your deployment runs on Alpine Linux (musl) and Deno is problematic,
-  // set YT_DLP_JS_RUNTIME=node explicitly.
-  if (runtimeIsInvalid) {
-    warnRuntimeOnce(
-      "invalid-runtime",
-      `[yt-dlp] Unsupported ${YT_DLP_JS_RUNTIME_ENV}="${rawRuntime}". Falling back to "deno".`
-    );
-  }
-
-  if (await isDenoAvailable()) {
-    return "deno";
-  }
-
-  if (runtime === "deno") {
-    warnRuntimeOnce(
-      "explicit-deno-unavailable",
-      '[yt-dlp] YT_DLP_JS_RUNTIME is set to "deno", but Deno runtime is unavailable. Falling back to "node". Install Deno or set YT_DLP_JS_RUNTIME=node.'
-    );
-    return "node";
-  }
-
-  if (runtimeIsInvalid) {
-    warnRuntimeOnce(
-      "invalid-runtime-deno-unavailable",
-      `[yt-dlp] YT_DLP_JS_RUNTIME="${rawRuntime}" is unsupported and Deno runtime is unavailable. Falling back to "node". Install Deno or set YT_DLP_JS_RUNTIME=node.`
-    );
-    return "node";
-  }
-
-  warnRuntimeOnce(
-    "default-deno-unavailable",
-    '[yt-dlp] Deno runtime is unavailable. Falling back to "node". Set YT_DLP_JS_RUNTIME=node to skip Deno checks.'
-  );
-  return "node";
-}
-
-async function appendYouTubeJsRuntimeArg(
-  args: string[],
-  url: string
-): Promise<void> {
-  if (!isYouTubeUrl(url)) {
-    return;
-  }
-  const runtimeFlag = await getYouTubeJsRuntimeFlag();
-  if (!runtimeFlag) {
-    return;
-  }
-  args.push(runtimeFlag, await getYouTubeJsRuntime());
-}
-
-function parseExtractorArgParts(value: unknown): string[] {
-  if (Array.isArray(value)) {
-    return value.flatMap((entry) => parseExtractorArgParts(entry));
-  }
-
-  if (typeof value !== "string") {
-    return [];
-  }
-
-  return value
-    .split(";")
-    .map((part) => part.trim())
-    .filter(Boolean);
-}
-
-function joinExtractorArgParts(parts: string[]): string | undefined {
-  const uniqueParts = Array.from(new Set(parts));
-  return uniqueParts.length > 0 ? uniqueParts.join(";") : undefined;
-}
-
-function withDefaultYouTubeExtractorArgs(
-  url: string,
-  flags: Record<string, any>
-): Record<string, any> {
-  if (!isYouTubeUrl(url)) {
-    return flags;
-  }
-
-  const providerScript = getProviderScript();
-  if (!providerScript) {
-    return flags;
-  }
-
-  const existingParts = parseExtractorArgParts(flags.extractorArgs);
-  const mergedParts = [...existingParts];
-
-  if (
-    !existingParts.some((part) => part.startsWith(YOUTUBE_PLAYER_CLIENT_ARG_PREFIX))
-  ) {
-    mergedParts.push(DEFAULT_YOUTUBE_PLAYER_CLIENT_EXTRACTOR_ARG);
-  }
-
-  const providerArg = `${PROVIDER_SCRIPT_ARG_PREFIX}${providerScript}`;
-  if (!existingParts.some((part) => part.startsWith(PROVIDER_SCRIPT_ARG_PREFIX))) {
-    mergedParts.push(providerArg);
-  }
-
-  const extractorArgs = joinExtractorArgParts(mergedParts);
-
-  return {
-    ...flags,
-    extractorArgs,
-  };
-}
-
-async function resolveYouTubeRemoteComponents(
-  url: string,
-  flags: Record<string, any>
-): Promise<Record<string, any>> {
-  if (!isYouTubeUrl(url) || !getProviderScript()) {
-    return flags;
-  }
-
-  const remoteComponentsDisabled =
-    flags.noRemoteComponents === true || flags.no_remote_components === true;
-  const explicitRemoteComponents =
-    flags.remoteComponents ?? flags.remote_components;
-  const remoteComponents = remoteComponentsDisabled
-    ? explicitRemoteComponents
-    : explicitRemoteComponents ?? DEFAULT_YOUTUBE_REMOTE_COMPONENTS;
-
-  if (remoteComponents === undefined) {
-    return flags;
-  }
-
-  if (!(await ytDlpSupportsRemoteComponents())) {
-    const {
-      remoteComponents: _remoteComponents,
-      remote_components: _remote_components,
-      ...rest
-    } = flags;
-    return rest;
-  }
-
-  return {
-    ...flags,
-    remoteComponents,
-  };
-}
-
-/**
- * Convert camelCase flag names to kebab-case CLI arguments
- */
-export function convertFlagToArg(flag: string): string {
-  return `--${flag.replace(/([A-Z])/g, "-$1").toLowerCase()}`;
-}
-
-// Map of short options to their long equivalents
-const SHORT_TO_LONG: Record<string, string> = {
-  f: "format",
-  S: "format-sort",
-  o: "output",
-  r: "limit-rate",
-  R: "retries",
-  N: "concurrent-fragments",
-  x: "extract-audio",
-  k: "keep-video",
-  j: "dump-json",
-  J: "dump-single-json",
-  "4": "force-ipv4",
-  "6": "force-ipv6",
-};
-
-/**
- * Convert flags object to yt-dlp CLI arguments array
- */
-export function flagsToArgs(flags: Record<string, any>): string[] {
-  const args: string[] = [];
-
-  for (const [key, value] of Object.entries(flags)) {
-    if (value === undefined || value === null) {
-      continue;
-    }
-
-    // Handle special cases
-    if (key === "extractorArgs") {
-      if (Array.isArray(value)) {
-        for (const extractorArg of value) {
-          if (extractorArg) {
-            args.push("--extractor-args", String(extractorArg));
-          }
-        }
-      } else if (typeof value === "string" || typeof value === "number") {
-        args.push("--extractor-args", String(value));
-      }
-      continue;
-    }
-
-    if (key === "addHeader") {
-      // addHeader is an array of "key:value" strings
-      if (Array.isArray(value)) {
-        for (const header of value) {
-          args.push("--add-header", header);
-        }
-      } else {
-        args.push("--add-header", value);
-      }
-      continue;
-    }
-
-    // Handle short options (single letter flags)
-    let argName: string;
-    if (SHORT_TO_LONG[key]) {
-      // Convert short option to long form
-      argName = `--${SHORT_TO_LONG[key]}`;
-    } else {
-      // Convert camelCase to kebab-case
-      argName = convertFlagToArg(key);
-    }
-
-    if (typeof value === "boolean") {
-      if (value) {
-        args.push(argName);
-      }
-    } else if (typeof value === "string" || typeof value === "number") {
-      args.push(argName, String(value));
-    } else if (Array.isArray(value)) {
-      // For arrays, join with comma or repeat the flag
-      args.push(argName, value.join(","));
-    }
-  }
-
-  return args;
-}
-
-/**
- * Execute yt-dlp with JSON output and return parsed result
- * @param url - Video URL
- * @param flags - yt-dlp flags
- * @param retryWithoutFormatRestrictions - If true, retry without format restrictions if format error occurs
- */
-export async function executeYtDlpJson(
-  rawUrl: string,
-  flags: Record<string, any> = {},
-  retryWithoutFormatRestrictions: boolean = true
-): Promise<any> {
-  await ensureYtDlpAvailable();
-  const url = preprocessUrl(rawUrl);
-  const partialFlags = withDefaultYouTubeExtractorArgs(url, flags);
-  const effectiveFlags = await resolveYouTubeRemoteComponents(url, partialFlags);
-  const { noWarnings: _noWarnings, ...jsonFlags } = effectiveFlags;
-  const args = [
-    "--dump-single-json",
-    "--no-warnings",
-    ...flagsToArgs(jsonFlags),
-  ];
-
-  // Add cookies after ensuring the file is in a yt-dlp-compatible format.
-  const cookiesPath = ensureCookiesFileIsNormalized();
-  if (cookiesPath) {
-    args.push("--cookies", cookiesPath);
-  }
-
-  await appendYouTubeJsRuntimeArg(args, url);
-
-  args.push(url);
-
-  const ytDlpPath = getConfiguredYtDlpPath();
-  console.log(`Executing: ${ytDlpPath} ${args.join(" ")}`);
-
-  return new Promise<any>((resolve, reject) => {
-    const subprocess = spawn(ytDlpPath, args, {
-      env: getYtDlpSpawnEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-
-    let stdout = "";
-    let stderr = "";
-
-    subprocess.stdout?.on("data", (data: Buffer) => {
-      stdout += data.toString();
-    });
-
-    subprocess.stderr?.on("data", (data: Buffer) => {
-      stderr += data.toString();
-    });
-
-    subprocess.on("close", async (code) => {
-      if (code !== 0) {
-        // Check if this is a format availability error
-        const isFormatError =
-          stderr.includes("Requested format is not available") ||
-          stderr.includes("format is not available") ||
-          stderr.includes("No video formats found");
-
-        // If it's a format error and we should retry, try again without format restrictions
-        if (isFormatError && retryWithoutFormatRestrictions) {
-          const hasFormatRestrictions =
-            (effectiveFlags.formatSort !== undefined &&
-              effectiveFlags.formatSort !== null) ||
-            (effectiveFlags.format !== undefined &&
-              effectiveFlags.format !== null) ||
-            (effectiveFlags.S !== undefined && effectiveFlags.S !== null) ||
-            (effectiveFlags.f !== undefined && effectiveFlags.f !== null);
-
-          if (hasFormatRestrictions) {
-            console.log(
-              "Format not available, retrying without format restrictions and with --ignore-config..."
-            );
-            try {
-              // Remove format-related flags
-              const retryFlags: Record<string, any> = {
-                ...effectiveFlags,
-                ignoreConfig: true,
-              };
-              delete retryFlags.formatSort;
-              delete retryFlags.format;
-              delete retryFlags.S;
-              delete retryFlags.f;
-              // Retry without format restrictions (don't retry again to avoid infinite loop)
-              const result = await executeYtDlpJson(url, retryFlags, false);
-              resolve(result);
-              return;
-            } catch (retryError) {
-              // If retry also fails, reject with original error
-              const error = new Error(
-                `yt-dlp process exited with code ${code}`
-              );
-              (error as any).stderr = stderr;
-              reject(error);
-              return;
-            }
-          } else if (!effectiveFlags.ignoreConfig) {
-            console.log(
-              "Format not available without explicit format flags, retrying with --ignore-config..."
-            );
-            try {
-              const retryFlags: Record<string, any> = {
-                ...effectiveFlags,
-                ignoreConfig: true,
-              };
-              // Retry without format restrictions (don't retry again to avoid infinite loop)
-              const result = await executeYtDlpJson(url, retryFlags, false);
-              resolve(result);
-              return;
-            } catch (retryError) {
-              // If retry also fails, reject with original error
-              const error = new Error(
-                `yt-dlp process exited with code ${code}`
-              );
-              (error as any).stderr = stderr;
-              reject(error);
-              return;
-            }
-          }
-        }
-
-        const error = new Error(`yt-dlp process exited with code ${code}`);
-        (error as any).stderr = stderr;
-        reject(error);
-        return;
-      }
-
-      if (
-        stderr &&
-        !stderr.includes("[download]") &&
-        !stderr.includes("[info]")
-      ) {
-        console.warn("yt-dlp stderr:", stderr);
-      }
-
-      try {
-        resolve(JSON.parse(stdout));
-      } catch (parseError) {
-        console.error("Failed to parse yt-dlp JSON output:", parseError);
-        console.error("Output:", stdout);
-        reject(new Error("Failed to parse yt-dlp output as JSON"));
-      }
-    });
-
-    subprocess.on("error", (error) => {
-      reject(error);
-    });
-  });
-}
-
-/**
- * Get channel URL from a video URL
- * Uses: yt-dlp <video_url> --print channel_url --skip-download
- */
-export async function getChannelUrlFromVideo(
-  videoUrlRaw: string,
-  networkConfig: Record<string, any> = {}
-): Promise<string | null> {
-  try {
-    await ensureYtDlpAvailable();
-  } catch (error) {
-    // Swallow availability errors: channel URL is supplementary metadata, not critical.
-    // Callers treat null as "not found", which is the correct fallback behavior here.
-    console.warn("yt-dlp unavailable when fetching channel URL:", error);
-    return null;
-  }
-
-  const videoUrl = preprocessUrl(videoUrlRaw);
-  const args = [
-    "--print",
-    "channel_url",
-    "--skip-download",
-    "--no-warnings",
-    ...flagsToArgs(networkConfig),
-  ];
-
-  // Add cookies after ensuring the file is in a yt-dlp-compatible format.
-  const cookiesPath = ensureCookiesFileIsNormalized();
-  if (cookiesPath) {
-    args.push("--cookies", cookiesPath);
-  }
-
-  await appendYouTubeJsRuntimeArg(args, videoUrl);
-
-  args.push(videoUrl);
-  const ytDlpPath = getConfiguredYtDlpPath();
-
-  return new Promise<string | null>((resolve, reject) => {
-    const subprocess = spawn(ytDlpPath, args, {
-      env: getYtDlpSpawnEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-
-    let stdout = "";
-    let stderr = "";
-
-    subprocess.stdout?.on("data", (data: Buffer) => {
-      stdout += data.toString();
-    });
-
-    subprocess.stderr?.on("data", (data: Buffer) => {
-      stderr += data.toString();
-    });
-
-    subprocess.on("close", (code) => {
-      if (code !== 0) {
-        console.warn(`Failed to get channel URL: ${stderr}`);
-        resolve(null);
-        return;
-      }
-
-      const channelUrl = stdout.trim();
-      resolve(channelUrl || null);
-    });
-
-    subprocess.on("error", (error) => {
-      console.warn(`Error getting channel URL:`, error);
-      resolve(null);
-    });
-  });
-}
-
-/**
- * Download channel avatar/thumbnail from channel URL
- * Uses: yt-dlp <channel_url> --write-thumbnail --playlist-items 0 -o <output_path>
- */
-export async function downloadChannelAvatar(
-  channelUrlRaw: string,
-  outputPath: string,
-  networkConfig: Record<string, any> = {}
-): Promise<boolean> {
-  try {
-    await ensureYtDlpAvailable();
-  } catch (error) {
-    // Swallow availability errors: avatar download failures are non-fatal.
-    // Callers treat false as "avatar unavailable" and continue without it.
-    console.warn("yt-dlp unavailable when downloading channel avatar:", error);
-    return false;
-  }
-
-  const channelUrl = preprocessUrl(channelUrlRaw);
-  const outputDir = path.dirname(outputPath);
-  const outputFilename = path.basename(outputPath, path.extname(outputPath));
-  const outputTemplate = resolveSafeChildPath(
-    outputDir,
-    `${outputFilename}.%(ext)s`
-  );
-
-  const args = [
-    "--write-thumbnail",
-    "--playlist-items",
-    "0",
-    "--skip-download",
-    "--no-warnings",
-    "--output",
-    outputTemplate,
-    ...flagsToArgs(networkConfig),
-  ];
-
-  // Add cookies after ensuring the file is in a yt-dlp-compatible format.
-  const cookiesPath = ensureCookiesFileIsNormalized();
-  if (cookiesPath) {
-    args.push("--cookies", cookiesPath);
-  }
-
-  await appendYouTubeJsRuntimeArg(args, channelUrl);
-
-  args.push(channelUrl);
-  const ytDlpPath = getConfiguredYtDlpPath();
-
-  return new Promise<boolean>((resolve, reject) => {
-    const subprocess = spawn(ytDlpPath, args, {
-      env: getYtDlpSpawnEnv(),
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-
-    let stderr = "";
-
-    subprocess.stderr?.on("data", (data: Buffer) => {
-      stderr += data.toString();
-    });
-
-    subprocess.on("close", (code) => {
-      if (code !== 0) {
-        console.warn(`Failed to download channel avatar: ${stderr}`);
-        // For Bilibili, this might be expected - log but don't fail completely
-        if (isBilibiliUrl(channelUrl)) {
-          console.warn(`Bilibili channel avatar download may not be supported by yt-dlp`);
-        }
-        resolve(false);
-        return;
-      }
-
-      // Check if the file was created (yt-dlp might save with different extension)
-      const possibleExtensions = ["jpg", "jpeg", "png", "webp"];
-      let foundFile = false;
-      for (const ext of possibleExtensions) {
-        const possiblePath = resolveSafeChildPath(
-          outputDir,
-          `${outputFilename}.${ext}`
-        );
-        if (pathExistsSafeSync(possiblePath, outputDir)) {
-          // If it's not a jpg, rename it to jpg
-          if (ext !== "jpg" && outputPath.endsWith(".jpg")) {
-            try {
-              moveSafeSync(possiblePath, outputDir, outputPath, outputDir, {
-                overwrite: true,
-              });
-            } catch (error) {
-              console.warn(`Failed to rename avatar to .jpg:`, error);
-              // If rename fails, just use the original file
-              resolve(true);
-              return;
-            }
-          } else if (ext !== "jpg") {
-            // File exists but with different extension - that's okay, we'll handle it
-            resolve(true);
-            return;
-          }
-          foundFile = true;
-          break;
-        }
-      }
-
-      // If no file found, check if outputPath exists (might have been created directly)
-      if (pathExistsSafeSync(outputPath, outputDir)) {
-        resolve(true);
-        return;
-      }
-
-      if (!foundFile) {
-        console.warn(`Channel avatar file not found after download. Checked extensions: ${possibleExtensions.join(", ")}`);
-        console.warn(`Output directory: ${outputDir}, Output filename: ${outputFilename}`);
-      }
-      resolve(foundFile);
-    });
-
-    subprocess.on("error", (error) => {
-      console.warn(`Error downloading channel avatar:`, error);
-      resolve(false);
-    });
-  });
-}
-
-/**
- * Execute yt-dlp with spawn for progress tracking.
- * Returns a subprocess-like object with kill() method.
- *
- * Uses PassThrough streams so the actual spawn can be deferred until
- * ensureYtDlpAvailable() confirms (and auto-installs) yt-dlp.
- */
-export function executeYtDlpSpawn(
-  rawUrl: string,
-  flags: Record<string, any> = {}
-): {
-  stdout: NodeJS.ReadableStream | null;
-  stderr: NodeJS.ReadableStream | null;
-  kill: (signal?: NodeJS.Signals) => boolean;
-  then: (
-    onFulfilled?: (value: void) => void | Promise<void>,
-    onRejected?: (reason: any) => void | Promise<void>
-  ) => Promise<void>;
-} {
-  const url = preprocessUrl(rawUrl);
-  const partialFlags = withDefaultYouTubeExtractorArgs(url, flags);
-
-  // PassThrough streams let callers attach handlers before the subprocess starts.
-  const stdoutPass = new PassThrough();
-  const stderrPass = new PassThrough();
-
-  let activeSubprocess: ReturnType<typeof spawn> | null = null;
-  let killRequested = false;
-  let killSignal: NodeJS.Signals | undefined;
-  let resolved = false;
-  let rejected = false;
-  let stderr = "";
-
-  const endPassThroughStreams = () => {
-    if (!stdoutPass.destroyed && !stdoutPass.readableEnded) {
-      stdoutPass.end();
-    }
-    if (!stderrPass.destroyed && !stderrPass.readableEnded) {
-      stderrPass.end();
-    }
-  };
-
-  const pipeOrForward = (
-    source: NodeJS.ReadableStream | null | undefined,
-    target: PassThrough
-  ) => {
-    if (!source) {
-      target.end();
-      return;
-    }
-
-    if (typeof (source as any).pipe === "function") {
-      source.pipe(target, { end: true });
-      return;
-    }
-
-    // Fallback used by tests that mock child streams as plain EventEmitters.
-    source.on("data", (chunk: Buffer | string) => target.write(chunk));
-    source.on("end", () => target.end());
-  };
-
-  // Accumulate stderr for error reporting (callers may also attach their own handlers).
-  stderrPass.on("data", (data: Buffer) => {
-    stderr += data.toString();
-  });
-
-  const promise = ensureYtDlpAvailable()
-    .then(
-      async () => {
-        const effectiveFlags = await resolveYouTubeRemoteComponents(
-          url,
-          partialFlags
-        );
-        const baseArgs = [...flagsToArgs(effectiveFlags)];
-
-        // Add cookies after ensuring the file is in a yt-dlp-compatible format.
-        const cookiesPath = ensureCookiesFileIsNormalized();
-        if (cookiesPath) {
-          baseArgs.push("--cookies", cookiesPath);
-        }
-
-        const ytDlpPath = getConfiguredYtDlpPath();
-        const args = [...baseArgs];
-        await appendYouTubeJsRuntimeArg(args, url);
-        args.push(url);
-
-        console.log(`Spawning: ${ytDlpPath} ${args.join(" ")}`);
-
-        return await new Promise<void>((resolve, reject) => {
-          if (killRequested) {
-            rejected = true;
-            endPassThroughStreams();
-            const error = new Error("yt-dlp process cancelled before start");
-            (error as any).code = "SIGTERM";
-            reject(error);
-            return;
-          }
-
-          activeSubprocess = spawn(ytDlpPath, args, {
-            env: getYtDlpSpawnEnv(),
-            stdio: ["ignore", "pipe", "pipe"],
-          });
-
-          pipeOrForward(activeSubprocess.stdout, stdoutPass);
-          pipeOrForward(activeSubprocess.stderr, stderrPass);
-
-          // Cancellation might happen between availability check and spawn.
-          if (killRequested && !activeSubprocess.killed) {
-            activeSubprocess.kill(killSignal);
-          }
-
-          activeSubprocess.on("close", (code, signal) => {
-            if (!resolved && !rejected) {
-              if (code === 0) {
-                resolved = true;
-                resolve();
-              } else if (killRequested) {
-                rejected = true;
-                const error = new Error(
-                  signal
-                    ? `yt-dlp process cancelled by ${signal}`
-                    : "yt-dlp process cancelled"
-                );
-                (error as any).stderr = stderr;
-                (error as any).code = signal || killSignal || "SIGTERM";
-                reject(error);
-              } else {
-                rejected = true;
-                const error = new Error(
-                  signal
-                    ? `yt-dlp process exited due to signal ${signal}`
-                    : `yt-dlp process exited with code ${code}`
-                );
-                (error as any).stderr = stderr;
-                (error as any).code = signal || code;
-                console.error("yt-dlp error output:", stderr);
-                reject(error);
-              }
-            }
-          });
-
-          activeSubprocess.on("error", (error) => {
-            if (!resolved && !rejected) {
-              rejected = true;
-              reject(error);
-            }
-          });
-        });
-      }
-    )
-    .catch((err) => {
-      // End streams without emitting stream "error" events that callers don't subscribe to.
-      endPassThroughStreams();
-      throw err;
-    });
-
-  return {
-    stdout: stdoutPass,
-    stderr: stderrPass,
-    kill: (signal?: NodeJS.Signals) => {
-      killRequested = true;
-      killSignal = signal;
-      if (activeSubprocess && !activeSubprocess.killed) {
-        return activeSubprocess.kill(signal);
-      }
-      // Allow cancelling before spawn starts.
-      return activeSubprocess === null ? true : false;
-    },
-    then: promise.then.bind(promise),
-  };
-}
-
-/**
- * Parse yt-dlp configuration text into flags object
- * Supports standard yt-dlp config file format (one option per line, # for comments)
- */
-export function parseYtDlpConfig(configText: string): Record<string, any> {
-  const flags: Record<string, any> = {};
-
-  if (!configText || typeof configText !== "string") {
-    return flags;
-  }
-
-  const lines = configText.split("\n");
-
-  for (const rawLine of lines) {
-    const line = rawLine.trim();
-
-    // Skip empty lines and comments
-    if (!line || line.startsWith("#")) {
-      continue;
-    }
-
-    // Parse the option
-    // Options can be:
-    // -f value
-    // --format value
-    // --some-flag (boolean)
-    // -x (short boolean)
-
-    let optionName: string | null = null;
-    let optionValue: string | boolean = true;
-
-    if (line.startsWith("--")) {
-      // Long option
-      const spaceIndex = line.indexOf(" ");
-      if (spaceIndex === -1) {
-        // Boolean flag (no value)
-        optionName = line.substring(2);
-      } else {
-        optionName = line.substring(2, spaceIndex);
-        optionValue = line.substring(spaceIndex + 1).trim();
-        // Remove surrounding quotes if present
-        if (
-          (optionValue.startsWith('"') && optionValue.endsWith('"')) ||
-          (optionValue.startsWith("'") && optionValue.endsWith("'"))
-        ) {
-          optionValue = optionValue.slice(1, -1);
-        }
-      }
-    } else if (line.startsWith("-") && !line.startsWith("--")) {
-      // Short option
-      const parts = line.split(/\s+/);
-      optionName = parts[0].substring(1);
-      if (parts.length > 1) {
-        optionValue = parts.slice(1).join(" ");
-        // Remove surrounding quotes if present
-        if (
-          typeof optionValue === "string" &&
-          ((optionValue.startsWith('"') && optionValue.endsWith('"')) ||
-            (optionValue.startsWith("'") && optionValue.endsWith("'")))
-        ) {
-          optionValue = optionValue.slice(1, -1);
-        }
-      }
-    }
-
-    if (optionName) {
-      // Convert kebab-case to camelCase for flags object
-      const camelCaseName = optionName.replace(/-([a-z])/g, (_, letter) =>
-        letter.toUpperCase()
-      );
-      flags[camelCaseName] = optionValue;
-    }
-  }
-
-  return flags;
-}
-
-/**
- * Get user's yt-dlp configuration from settings
- * @param url - Optional URL to contextually filter settings (e.g. proxy only for YouTube)
- */
-export function getUserYtDlpConfig(url?: string): Record<string, any> {
-  try {
-    if (!isAdminTrustLevelAtLeast("container")) {
-      return {};
-    }
-
-    const settings = storageService.getSettings();
-    const configText = settings.ytDlpConfig;
-    const proxyOnlyYoutube = settings.proxyOnlyYoutube === true;
-
-    if (configText) {
-      const parsedConfig = parseYtDlpConfig(configText);
-      console.log("Parsed user yt-dlp config:", parsedConfig);
-
-      // If proxy is restricted to YouTube only, and we have a non-YouTube URL
-      if (proxyOnlyYoutube && url) {
-        const isYoutube = isYouTubeUrl(url);
-        if (!isYoutube) {
-          console.log(
-            "Proxy restricted to YouTube only. Removing proxy settings for:",
-            url
-          );
-          // Remove proxy-related settings
-          delete parsedConfig.proxy;
-          // Also remove potentially related network options if they are usually proxy-specific?
-          // sticking to just 'proxy' as per request and standard usage.
-        }
-      }
-
-      return parsedConfig;
-    }
-  } catch (error) {
-    console.error("Error reading user yt-dlp config:", error);
-  }
-  return {};
-}
-
-/**
- * Extract network-related options from user config
- * These are safe to apply to all operations (search, info, download)
- */
-export function getNetworkConfigFromUserConfig(
-  userConfig: Record<string, any>
-): Record<string, any> {
-  const networkOptions: Record<string, any> = {};
-
-  // Proxy settings
-  if (userConfig.proxy) {
-    networkOptions.proxy = userConfig.proxy;
-  }
-
-  // Rate limiting
-  if (userConfig.r || userConfig.limitRate) {
-    networkOptions.limitRate = userConfig.r || userConfig.limitRate;
-  }
-
-  // Socket timeout
-  if (userConfig.socketTimeout) {
-    networkOptions.socketTimeout = userConfig.socketTimeout;
-  }
-
-  // Force IPv4/IPv6
-  if (userConfig.forceIpv4 || userConfig["4"]) {
-    networkOptions.forceIpv4 = true;
-  }
-  if (userConfig.forceIpv6 || userConfig["6"]) {
-    networkOptions.forceIpv6 = true;
-  }
-
-  // Geo bypass
-  if (userConfig.xff) {
-    networkOptions.xff = userConfig.xff;
-  }
-
-  // Sleep/rate limiting
-  if (userConfig.sleepRequests) {
-    networkOptions.sleepRequests = userConfig.sleepRequests;
-  }
-  if (userConfig.sleepInterval || userConfig.minSleepInterval) {
-    networkOptions.sleepInterval =
-      userConfig.sleepInterval || userConfig.minSleepInterval;
-  }
-  if (userConfig.maxSleepInterval) {
-    networkOptions.maxSleepInterval = userConfig.maxSleepInterval;
-  }
-
-  // Retries
-  if (userConfig.retries || userConfig.R) {
-    networkOptions.retries = userConfig.retries || userConfig.R;
-  }
-
-  return networkOptions;
-}
-
-/**
- * Error thrown when proxy configuration is invalid
- */
-export class InvalidProxyError extends Error {
-  readonly proxyUrl: string;
-  readonly originalError?: Error;
-
-  constructor(proxyUrl: string, originalError?: Error) {
-    super(`Invalid proxy URL: ${proxyUrl}`);
-    this.name = "InvalidProxyError";
-    this.proxyUrl = proxyUrl;
-    this.originalError = originalError;
-  }
-}
-
-/**
- * Helper to convert a proxy URL string into an Axios config object
- * Supports http/https/socks5 proxies with authentication
- * Format: http://user:pass@host:port or socks5://user:pass@host:port
- *
- * @throws {InvalidProxyError} If the proxy URL is malformed - this prevents
- *         silent fallback to direct connection which could expose user's real IP
- */
-export function getAxiosProxyConfig(proxyUrl: string): any {
-  if (!proxyUrl) return {};
-
-  try {
-    const url = new URL(proxyUrl);
-    const protocol = url.protocol.replace(":", "");
-
-    // Validate that we have a hostname
-    if (!url.hostname) {
-      throw new InvalidProxyError(proxyUrl, new Error("Missing hostname"));
-    }
-
-    // Check if this is a SOCKS proxy
-    if (protocol.startsWith("socks")) {
-      // Force remote DNS resolution by using socks5h for all socks5 variants
-      // socks-proxy-agent uses socks5h to indicate remote DNS resolution
-      // Remote DNS resolution helps avoid:
-      // - DNS pollution/poisoning on the client side
-      // - DNS blocking by local network
-      // - DNS resolution delays
-      let agentUrl = proxyUrl;
-
-      // Convert socks5 to socks5h (remote DNS resolution)
-      // If user already specified socks5h, keep it as-is
-      if (protocol === "socks5") {
-        agentUrl = proxyUrl.replace("socks5://", "socks5h://");
-        console.log("Converted socks5 to socks5h for remote DNS resolution");
-      } else if (protocol === "socks5h") {
-        // Already using remote DNS, no conversion needed
-        console.log("Using socks5h (remote DNS resolution)");
-      }
-      // Note: socks4/socks4a don't support remote DNS, so we leave them as-is
-
-      // Use SocksProxyAgent for SOCKS proxy support
-      const agent = new SocksProxyAgent(agentUrl);
-      return {
-        httpAgent: agent,
-        httpsAgent: agent,
-        proxy: false, // Disable axios built-in proxy when using custom agents
-      };
-    }
-
-    // Validate protocol for non-SOCKS proxies
-    if (protocol !== "http" && protocol !== "https") {
-      throw new InvalidProxyError(
-        proxyUrl,
-        new Error(`Unsupported proxy protocol: ${protocol}`)
-      );
-    }
-
-    // Handle HTTP/HTTPS proxies
-    const isHttps = protocol === "https";
-    const defaultPort = isHttps ? 443 : 80;
-
-    // Axios proxy config structure
-    const proxyConfig: any = {
-      protocol: protocol,
-      host: url.hostname,
-      port: parseInt(url.port, 10) || defaultPort,
-    };
-
-    if (url.username || url.password) {
-      proxyConfig.auth = {
-        username: decodeURIComponent(url.username),
-        password: decodeURIComponent(url.password),
-      };
-    }
-
-    return { proxy: proxyConfig };
-  } catch (error) {
-    // Re-throw InvalidProxyError as-is
-    if (error instanceof InvalidProxyError) {
-      throw error;
-    }
-    // Wrap other errors (like URL parsing errors) in InvalidProxyError
-    // This ensures we fail rather than silently falling back to direct connection
-    console.error("Invalid proxy URL:", proxyUrl, error);
-    throw new InvalidProxyError(
-      proxyUrl,
-      error instanceof Error ? error : new Error(String(error))
-    );
-  }
+  resetYtDlpAvailablePromise();
+  resetProviderPluginCache();
+  resetResolvedYtDlpPath();
+  resetRuntimeCaches();
+  resetCookiesFileCache();
 }
